% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{format_school_year}
\alias{format_school_year}
\title{Format end year as school year string}
\usage{
format_school_year(end_year)
}
\arguments{
\item{end_year}{Integer end year (e.g., 2024)}
}
\value{
Character string like "2023-24"
}
\description{
Converts 2024 to "2023-24" format
}
\keyword{internal}
