% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_raw_enrollment.R
\name{get_raw_enr}
\alias{get_raw_enr}
\title{Download raw enrollment data from OSPI}
\usage{
get_raw_enr(end_year)
}
\arguments{
\item{end_year}{School year end (2024 = 2023-24 school year)}
}
\value{
Data frame with raw enrollment data
}
\description{
Downloads enrollment data from Washington State Report Card on data.wa.gov
using the Socrata Open Data API (SODA).
}
\keyword{internal}
