% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{merge_eds_data}
\alias{merge_eds_data}
\title{Merge EDS data into raw directory output}
\usage{
merge_eds_data(entity_data, eds_data)
}
\arguments{
\item{entity_data}{Raw entity data}

\item{eds_data}{EDS directory data}
}
\value{
Merged data frame
}
\description{
Simple merge for raw (tidy=FALSE) output.
}
\keyword{internal}
