% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{parse_school_year}
\alias{parse_school_year}
\title{Convert school year string to end year}
\usage{
parse_school_year(school_year)
}
\arguments{
\item{school_year}{Character string like "2023-24" or "2023-2024"}
}
\value{
Integer end year
}
\description{
Converts "2023-24" format to 2024 (end year)
}
\keyword{internal}
