% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fetch_directory.R
\name{read_cache_directory}
\alias{read_cache_directory}
\title{Read directory data from cache}
\usage{
read_cache_directory(end_year, cache_type)
}
\arguments{
\item{end_year}{School year end}

\item{cache_type}{Type of cache}
}
\value{
Cached data frame
}
\description{
Read directory data from cache
}
\keyword{internal}
